camera/requirements.txt
aioboto3
python-dotenv
setproctitle